feat: impl NoRegisterClubForProfessor
